multi_modal_model/requirements.txt
numpy
pandas
scikit-learn
torch
torchvision
open_clip_torch
transformers
Pillow
